chore:(review) merged files of txn and proc
